chore: Sheet worker code syntax cleanup (#12)
This change makes a few cosmetic changes to the sheet worker script code in the character sheet. These are mostly made to make the code more readableand maintainable, A few functional changes were added for future use, but these should have no impact on the operation of the character sheet until they are hooked up to other events. The following changes are being made:

* Removed separate `<script>` tags in keeping with Roll20 best practices guidance.
* Updated anonymous functions to use fat-arrow (`=>`) notation instead of using the `function` keyword. This serves as a visual cue for developers that the function in question is not intended to be used outside of its current context.
* Added support for Promises and async/await code. This is done by including the [community code](https://github.com/onyxring/Roll20Async) for this purpose.
* Added wrapper function for the [Fetch API](https://developer.mozilla.org/en-US/docs/Web/API/Fetch_API) to reduce duplicated code in future character sheet feature implementations requiring the API.
* Linting code to conform to readability standards.

Co-authored-by: Jim Evans <james.evans@salesforce.com>